New poll command and bug fixes
Made a new poll command that takes a prompt and options from a user, then sends a message with the prompt and emojis with options, reacting to the message so users can vote/poll.
Had some bugs where the logging had some issues in the paint() and teams() functions. It's because I didn't use the logger() function that I made.