Bugfixes
The search wasn't working properly because i forgot to put spaces in between the words. Meaning it would search 'iamsaraciki' instead of 'i am sara ciki'.
Also forgot to remove the path to ffmpeg on my PC on this doc.
The server im running it on had problems because it was running redhat 7.9 and its a huge pain to try and install ffmpeg version 4 on it. I swapped to an ubuntu server, and it is so much easier. Had a much easier time setting everything up too. Now all the problems with the search function should be fixed.