Added remove(range), shuffle and skipto
Can now remove a range of songs from the queue, e.g. `Kevin remove 3 6` which will remove songs from number 3 to 6, inclusively.
Shuffle will shuffle the order of songs in the queue.
Skipto lets you skip to a song in the queue, removing every song before it.

There's an issue with the manual skipping, where if the auto-skip triggers during the manual skip, it'll raise an already_playing error and remove an additional song from the queue instead of just one. Trying to figure out how to fix that.