Bugfix with autosuggest (autocorrect)
The autosuggest was enabled by default and caused some wiki searches (like dua lipa) to be autocorrected into something else (due lips). Fixed that by turning that off and also added better error handling messages for disambiguation error and added an alias for the wikisearch so I can just call it with search.